GeoPackage: GetNextArrowArray(): report enum coded field domain as Arrow dictionary
