+++ ogr/ogrsf_frmts/generic/makefile.vc
@@ -4,7 +4,8 @@ OBJ	=	ogrsfdriverregistrar.obj ogrlayer.obj ogr_gensql.obj \
 		ogr_attrind.obj ogr_miattrind.obj ogrlayerdecorator.obj \
 		ogrwarpedlayer.obj ogrunionlayer.obj ogrlayerpool.obj \
 		ogrmutexedlayer.obj ogrmutexeddatasource.obj \
-		ogremulatedtransaction.obj ogreditablelayer.obj
+		ogremulatedtransaction.obj ogreditablelayer.obj \
+		ograrrowarrayhelper.obj
 
 
 GDAL_ROOT	=	..\..\..
