+++ ogr/ogrsf_frmts/generic/ograrrowarrayhelper.h
@@ -35,8 +35,11 @@
 #include "ogrsf_frmts.h"
 #include "ogr_recordbatch.h"
 
-class OGRArrowArrayHelper
+class CPL_DLL OGRArrowArrayHelper
 {
+    OGRArrowArrayHelper(const OGRArrowArrayHelper&) = delete;
+    OGRArrowArrayHelper& operator= (const OGRArrowArrayHelper&) = delete;
+
 public:
     bool bIncludeFID = false;
     int nMaxBatchSize = 0;
@@ -50,220 +53,10 @@ public:
     int64_t* panFIDValues = nullptr;
     struct ArrowArray* m_out_array = nullptr;
 
-    OGRArrowArrayHelper(OGRFeatureDefn* poFeatureDefn,
+    OGRArrowArrayHelper(GDALDataset* poDS,
+                        OGRFeatureDefn* poFeatureDefn,
                         const CPLStringList& aosArrowArrayStreamOptions,
-                        struct ArrowArray* out_array):
-        bIncludeFID(CPLTestBool(
-            aosArrowArrayStreamOptions.FetchNameValueDef("INCLUDE_FID", "YES"))),
-        nMaxBatchSize(atoi(
-            aosArrowArrayStreamOptions.FetchNameValueDef("MAX_FEATURES_IN_BATCH", "65536"))),
-        m_out_array(out_array)
-    {
-        memset(out_array, 0, sizeof(*out_array));
-
-        if( nMaxBatchSize <= 0 )
-            nMaxBatchSize = 1;
-        if( nMaxBatchSize > INT_MAX - 1 )
-            nMaxBatchSize = INT_MAX - 1;
-
-        nFieldCount = poFeatureDefn->GetFieldCount();
-        nGeomFieldCount = poFeatureDefn->GetGeomFieldCount();
-        mapOGRFieldToArrowField.resize(nFieldCount, -1);
-        mapOGRGeomFieldToArrowField.resize(nGeomFieldCount, -1);
-        abNullableFields.resize(nFieldCount);
-
-        if( bIncludeFID )
-        {
-            nChildren ++;
-        }
-        for(int i = 0; i < nFieldCount; i++ )
-        {
-            const auto poFieldDefn = poFeatureDefn->GetFieldDefn(i);
-            abNullableFields[i] = poFieldDefn->IsNullable();
-            if( !poFieldDefn->IsIgnored() )
-            {
-                mapOGRFieldToArrowField[i] = nChildren;
-                nChildren ++;
-            }
-        }
-        for(int i = 0; i < nGeomFieldCount; i++ )
-        {
-            if( !poFeatureDefn->GetGeomFieldDefn(i)->IsIgnored() )
-            {
-                mapOGRGeomFieldToArrowField[i] = nChildren;
-                nChildren ++;
-            }
-        }
-
-        anArrowFieldMaxAlloc.resize(nChildren);
-
-        out_array->release = OGRLayer::ReleaseArray;
-
-        out_array->length = nMaxBatchSize;
-        out_array->null_count = -1;
-
-        out_array->n_children = nChildren;
-        out_array->children = static_cast<struct ArrowArray**>(
-                            CPLCalloc(nChildren, sizeof(struct ArrowArray*)));
-        out_array->release = OGRLayer::ReleaseArray;
-        out_array->n_buffers = 1;
-        out_array->buffers = static_cast<const void**>(CPLCalloc(1, sizeof(void*)));
-
-        // Allocate buffers
-
-        if( bIncludeFID )
-        {
-            out_array->children[0]= static_cast<struct ArrowArray*>(
-                CPLCalloc(1, sizeof(struct ArrowArray)));
-            auto psChild = out_array->children[0];
-            psChild->release = OGRLayer::ReleaseArray;
-            psChild->length = nMaxBatchSize;
-            psChild->n_buffers = 2;
-            psChild->buffers = static_cast<const void**>(CPLCalloc(2, sizeof(void*)));
-            panFIDValues = static_cast<int64_t*>(
-                VSI_MALLOC_ALIGNED_AUTO_VERBOSE(sizeof(int64_t) * nMaxBatchSize));
-            if( panFIDValues == nullptr )
-                goto error;
-            psChild->buffers[1] = panFIDValues;
-        }
-
-        for(int i = 0; i < nFieldCount; i++ )
-        {
-            const int iArrowField = mapOGRFieldToArrowField[i];
-            if( iArrowField >= 0 )
-            {
-                const auto poFieldDefn = poFeatureDefn->GetFieldDefn(i);
-                out_array->children[iArrowField]= static_cast<struct ArrowArray*>(
-                    CPLCalloc(1, sizeof(struct ArrowArray)));
-                auto psChild = out_array->children[iArrowField];
-
-                psChild->release = OGRLayer::ReleaseArray;
-                psChild->length = nMaxBatchSize;
-                const auto eSubType = poFieldDefn->GetSubType();
-                size_t nEltSize = 0;
-                switch( poFieldDefn->GetType() )
-                {
-                    case OFTInteger:
-                    {
-                        if( eSubType == OFSTBoolean )
-                        {
-                            nEltSize = sizeof(uint8_t);
-                        }
-                        else if( eSubType == OFSTInt16 )
-                        {
-                            nEltSize = sizeof(int16_t);
-                        }
-                        else
-                        {
-                            nEltSize = sizeof(int32_t);
-                        }
-                        break;
-                    }
-                    case OFTInteger64:
-                    {
-                        nEltSize = sizeof(int64_t);
-                        break;
-                    }
-                    case OFTReal:
-                    {
-                        if( eSubType == OFSTFloat32 )
-                        {
-                            nEltSize = sizeof(float);
-                        }
-                        else
-                        {
-                            nEltSize = sizeof(double);
-                        }
-                        break;
-                    }
-                    case OFTString:
-                    case OFTBinary:
-                    {
-                        psChild->n_buffers = 3;
-                        psChild->buffers = static_cast<const void**>(CPLCalloc(3, sizeof(void*)));
-                        psChild->buffers[1] =
-                            VSI_MALLOC_ALIGNED_AUTO_VERBOSE(sizeof(uint32_t) * (1 + nMaxBatchSize));
-                        if( psChild->buffers[1] == nullptr )
-                            goto error;
-                        memset( (void*)psChild->buffers[1], 0, sizeof(uint32_t) * (1 + nMaxBatchSize));
-                        constexpr size_t DEFAULT_STRING_SIZE = 10;
-                        anArrowFieldMaxAlloc[iArrowField] = DEFAULT_STRING_SIZE * nMaxBatchSize;
-                        psChild->buffers[2] =
-                            VSI_MALLOC_ALIGNED_AUTO_VERBOSE(anArrowFieldMaxAlloc[iArrowField]);
-                        if( psChild->buffers[2] == nullptr )
-                            goto error;
-                        break;
-                    }
-
-                    case OFTDate:
-                    {
-                        nEltSize = sizeof(int32_t);
-                        break;
-                    }
-
-                    case OFTTime:
-                    {
-                        nEltSize = sizeof(int32_t);
-                        break;
-                    }
-
-                    case OFTDateTime:
-                    {
-                        nEltSize = sizeof(int64_t);
-                        break;
-                    }
-
-                    default:
-                        break;
-                }
-
-                if( nEltSize != 0 )
-                {
-                    psChild->n_buffers = 2;
-                    psChild->buffers = static_cast<const void**>(CPLCalloc(2, sizeof(void*)));
-                    psChild->buffers[1] =
-                        VSI_MALLOC_ALIGNED_AUTO_VERBOSE(nEltSize * nMaxBatchSize);
-                    if( psChild->buffers[1] == nullptr )
-                        goto error;
-                    memset( const_cast<void*>(psChild->buffers[1]), 0, nEltSize * nMaxBatchSize );
-                }
-            }
-        }
-
-        for(int i = 0; i < nGeomFieldCount; i++ )
-        {
-            const int iArrowField = mapOGRGeomFieldToArrowField[i];
-            if( iArrowField >= 0 )
-            {
-                out_array->children[iArrowField]= static_cast<struct ArrowArray*>(
-                    CPLCalloc(1, sizeof(struct ArrowArray)));
-                auto psChild = out_array->children[iArrowField];
-
-                psChild->release = OGRLayer::ReleaseArray;
-                psChild->length = nMaxBatchSize;
-
-                psChild->n_buffers = 3;
-                psChild->buffers = static_cast<const void**>(CPLCalloc(3, sizeof(void*)));
-                psChild->buffers[1] =
-                    VSI_MALLOC_ALIGNED_AUTO_VERBOSE(sizeof(uint32_t) * (1 + nMaxBatchSize));
-                if( psChild->buffers[1] == nullptr )
-                    goto error;
-                memset((void*)psChild->buffers[1], 0, sizeof(uint32_t) * (1 + nMaxBatchSize));
-                constexpr size_t DEFAULT_WKB_SIZE = 100;
-                anArrowFieldMaxAlloc[iArrowField] = DEFAULT_WKB_SIZE * nMaxBatchSize;
-                psChild->buffers[2] =
-                    VSI_MALLOC_ALIGNED_AUTO_VERBOSE(anArrowFieldMaxAlloc[iArrowField]);
-                if( psChild->buffers[2] == nullptr )
-                    goto error;
-            }
-        }
-
-        return;
-
-error:
-        out_array->release(out_array);
-        memset(out_array, 0, sizeof(*out_array));
-    }
+                        struct ArrowArray* out_array);
 
     bool SetNull(int iArrowField, int iFeat)
     {
@@ -432,6 +225,9 @@ error:
         m_out_array->release(m_out_array);
         memset(m_out_array, 0, sizeof(*m_out_array));
     }
+
+    static bool FillDict(struct ArrowArray* psChild,
+                         const OGRCodedFieldDomain* poCodedDomain);
 
 };
